+++ src/main/java/si/matjazcerkvenik/alertmonitor/util/AmProps.java
@@ -85,7 +85,7 @@ public class AmProps {
         if (DEV_ENV) {
             ALERTMONITOR_PROMETHEUS_SERVER = "https://elasticvm/prometheus";
             ALERTMONITOR_MONGODB_ENABLED = true;
-            ALERTMONITOR_MONGODB_CONNECTION_STRING = "mongodb://admin:mongodbpassword@elasticvm:27017/?authSource=admin";
+            ALERTMONITOR_MONGODB_CONNECTION_STRING = "mongodb://admin:mongodbpassword@lionvm:27017/?authSource=admin";
             ALERTMONITOR_MONGODB_DB_NAME = "alertmonitor-dev";
 //            ALERTMONITOR_DATAPROVIDERS_CONFIG_FILE = "providers.yml";
             ALERTMONITOR_DATAPROVIDERS_CONFIG_FILE = "/Users/matjaz/Library/CloudStorage/Dropbox/monis/config/alertmonitor-providers.yml";