+++ src/main/java/si/matjazcerkvenik/alertmonitor/util/LogFactory.java
@@ -1,3 +1,18 @@
+/*
+   Copyright 2021 Matjaž Cerkvenik
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+ */
 package si.matjazcerkvenik.alertmonitor.util;
 
 import si.matjazcerkvenik.simplelogger.SimpleLogger;
@@ -23,20 +38,5 @@ public class LogFactory {
         }
         return logger;
     }
-
-    public static SimpleLogger getAlertLog() {
-        if (alertLog == null) {
-            if (AmProps.DEV_ENV) {
-                // write file in local working directory
-                alertLog = new SimpleLogger("./alerts.log");
-            } else {
-                // in production environment (aka when running inside container)
-                alertLog = new SimpleLogger("/opt/alertmonitor/log/alerts.log");
-            }
-            alertLog.setVerbose(false);
-        }
-        return alertLog;
-    }
-
 
 }