+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.4.6-SNAPSHOT
 
 Working on timeline
+* [CHANGE] Removed separate alert.log (introduced in 2.3.0)
 
 ## 2.4.5-SNAPSHOT